entre maquinas y tintas/Entre-Maquinas-y-Tintas/requirements.txt
Django==4.2
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2==2.9.6
django-cors-headers==3.14.0